ingestion/stream/requirements.txt
websocket-client>=1.7.0
kafka-python>=2.0.2
python-dotenv>=1.0.0
boto3>=1.34.0
pyyaml>=6.0
pandas>=2.0.0
pyarrow>=14.0.0
